feat: split expired hashmap and utils functions. Add command_utils crate
