+++ src/commands/command_lib/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "rus_key_command_lib"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
