+++ src/commands/expired_commands/src/const.rs
@@ -1,0 +1,1 @@
+pub const EXPIRED: &str = "expired";
