+++ src/commands/hashmap_commands/src/lib.rs
@@ -1,1 +1,3 @@
 pub mod hashmap;
+mod hgetall;
+mod hmset;
