+++ src/commands/string_commands/src/get.rs
@@ -1,7 +1,7 @@
 use std::str::SplitAsciiWhitespace;
 use expired_commands::expired::get_key_expired;
 use rus_key_db::db::{DataType, Db};
-use crate::utils::get_parts;
+use rus_key_command_lib::get_parts;
 use crate::r#const::EMPTY;
 
 pub fn get(
