+++ src/commands/string_commands/src/set.rs
@@ -1,9 +1,10 @@
 use std::str::SplitAsciiWhitespace;
 use expired_commands::expired::{ExpiredCommand, get_key_expired};
 use rus_key_db::db::{DataType, Db};
+use rus_key_command_lib::get_parts;
+use crate::utils::general_command;
 use crate::get::get;
 use crate::r#const::{ExtraArgs, SetError};
-use crate::utils::{general_command, get_parts};
 
 pub fn set(parts: &mut SplitAsciiWhitespace, db: &mut Db) -> String {
     let (key, value) = get_parts(parts, true);
