+++ src/commands/string_commands/src/string.rs
@@ -1,7 +1,7 @@
 use rus_key_trait::command_trait::Command;
 use rus_key_db::db::Db;
 use crate::append::append;
-use crate::utils::handle_accumulation;
+use crate::handle_accumulation::handle_accumulation;
 use crate::incrby_float::incrby_float;
 use crate::get::get;
 use crate::get_del::get_del;
