+++ src/commands/utils_commands/src/lib.rs
@@ -1,1 +1,6 @@
 pub mod utils;
+mod check_expired;
+mod rename;
+mod randomkey;
+mod check_type;
+mod del_key;
